Schema Inheritance
Made the Article schema the base schema for all articles and made the books and wine schemas inherit the fields from Article.
Made a single endpoint for getting articles of whatever kind you pass to the API.
Updated the ajax call in the books page to use the new endpoint for getting books articles.