Todo List
Tackled some items off the to do list.
Updated the email on the footer.
Removed the words social media links on the footer.
Updated the thumbnail photos on the home screen and changed how they are layed out.
Removed the word articles on the article list pages.
Removed the word Authors on the book article page.
Added space below the next/previous article navigation.