Update Article Work
Started working on getting update articles to work.
Found out there are issues with the preview article component.
Cannot set props of the given article and book article cant find authors array.
Need to make a single form component for update and create article and have sub components that differ based on which view we are in.